Main Page Edit (#2)
* Add files via upload

Initial rough draft website


* Delete __pycache__/

* Copied new code over existing code. Updated  main/home page. Added Pet Database

---------
