--- petFeederProject/static/styles.css
@@ -1,225 +1,0 @@
-*,
-*::before,
-*::after {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-:root {
-  /* Main colors */
-  --primary-background: #f5f5f5;    /* Light gray background */
-  --navbar-bg: #e5f3fd;             /* Deep charcoal for navbar */
-  --profile-bg: #2a9df4;            /* Lighter charcoal for profile */
-  --login-bg: #e5f3fd;
-  /* Accent colors */
-  --accent: #ff7e67;                /* Coral accent */
-  --text-light: #ffffff;            /* White text */
-  --text-dark: #2d3436;             /* Dark text */
-}
-
-/* Replace opacity with rgba() for background colors */
-body{
-  min-height: 100vh;
-  margin: 0;
-  overflow-x: hidden;
-}
-
-.container {
-  background-color: var(--primary-background);
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  position: relative;
-  overflow-y: auto;
-}
-
-
-/* ***start of profile-page*** */
-.user-info-container {
-  background-color: var(--profile-bg);
-  color: var(--text-dark); /* Instead of #001972 with opacity */
-  display: flex;
-  border-radius: 0px 70px 25px 0px;
-  max-width: fit-content;
-  width: 40%;
-  min-height: fit-content;
-  text-align: center; 
-}
-
-.user-info {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 10px;
-  margin: 10px;
-}
-
-.user-info h1 {
-  color: var(--text-dark);
-  margin: 10px 0;
-}
-
-.user-info img {
-  width: 128px;
-  height: 128px;
-  border-radius: 50%;
-  margin: 5px;
-  outline: var(--text-light) solid 2px;  
-}
-
-
-.user-info p {
-  background-color: var(--text-dark);
-  margin: 5px;
-  padding: 5px;
-  border-radius: 25px;
-  background-color: var(--login-bg);
-  justify-content: center;
-  outline: var(--text-light) solid 2px; 
-}
-
-.user-info a {
-  color: var(--text-dark);
-  margin: 10px;
-  border-radius: 25px;  
-}
-/* ***end of profile-page*** */
-
-
-.header {
-  background-color: var(--navbar-bg);
-  color: var(--text-light);
-  text-align: center;
-  min-width: fit-content;
-  font-size: 3vh 3vw;
-}
-
-.navbar {
-  background-color: var(--navbar-bg);
-  color: var(--text-dark); /* Instead of opacity property */
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 5px 20px;
-}
-
-.navbar a {
-  color: var(--text-dark);
-  text-decoration: none;
-  margin-right: 15px;
-  font-size: 1.05em;
-}
-
-.navbar a:hover {
-  color: var(--accent);
-}
-
-/* Profile picture and username container */
-.user-profile {
-  display: flex;
-  align-items: center; /* Aligns the profile picture and username vertically */
-}
-
-/* Styling for the profile picture */
-.profile-img {
-  width: 40px; /* Adjust the size of the profile picture */
-  height: 40px;
-  border-radius: 50%; /* Makes the image round */
-  margin-right: 10px; /* Adds space between the profile picture and the username */
-}
-
-/* Styling for the username */
-.username {
-  color: rgba(221, 221, 221, 1);
-  font-weight: bold;
-  margin-right: 15px; /* Space between username and logout link */
-}
-
-/* Styling for the logout link */
-.logout {
-  color: rgba(255, 99, 71, 1);
-  text-decoration: none;
-  font-size: 1rem;
-}
-
-.logout:hover {
-  text-decoration: underline;
-}
-
-/* Styling for the login link */
-.login-link {
-  color: rgba(221, 221, 221, 1)
-  ;
-  text-decoration: none;
-}
-
-/* Profile info section */
-.profile-info {
-  display: flex;
-  align-items: center; /* Vertically centers the profile picture and username */
-}
-
-/* ***Start of Login-Page*** */
-
-
-
-.form-container {
-  background-color: #e5f3fd;
-  padding: 2rem;
-  border-radius: 10px;
-  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-  width: 90%;
-  max-width: 400px;
-  
-  /* Centering with position and transform */
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.form-container h1 {
-  text-align: center;
-  color: #2a9df4;
-  margin-bottom: 1rem;
-}
-
-.form-control {
-  width: 100%;
-  padding: 0.8rem;
-  margin: 0.5rem 0;
-  border: 1px solid #d0e6f7;
-  border-radius: 5px;
-  font-size: 1rem;
-}
-
-.form-control:focus {
-  outline: none;
-  border-color: #2a9df4;
-  box-shadow: 0 0 5px rgba(42, 157, 244, 0.3);
-}
-
-.login-button {
-  text-align: center;
-}
-
-button#submit {
-  background-color: #a1e3a1; /* Light green background */
-  color: #ffffff; /* White text */
-  padding: 0.8rem 1.5rem;
-  border: none;
-  border-radius: 25px; /* Rounded shape */
-  font-size: 1rem;
-  cursor: pointer;
-  transition: background-color 0.3s, transform 0.2s;
-  width: 100%;
-  font-weight: bold;
-}
-
-button#submit:hover {
-  background-color: #87d387; /* Darker shade on hover */
-  transform: scale(1.05); /* Slightly enlarges the button on hover */
-}
-
-
