--- petFeederProject/templates/edit_profile.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Edit page</title>
-</head>
-<body>
-    <form method="POST" enctype="multipart/form-data">
-        Bio: <textarea name="bio">{{ user[2] }}</textarea><br>
-        Profile Picture: <input type="file" name="profile_picture"><br>
-        <button type="submit">Update Profile</button>
-    </form>
-</body>
-</html>
