+++ templates/login.html
@@ -1,0 +1,46 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="{{ url_for('static', filename='styles.css') }}" rel="stylesheet">
+
+    <title>User Login Page</title>
+</head>
+<body>
+    <div class="container">
+        <div class="form-container">
+            <div class="login-form">
+                <h1 id="welcome-notice">Welcome!</h1>
+                <form  method="POST" action="{{ url_for('login') }}">
+                    <input
+                    required
+                    name="username"
+                    id="username"
+                    type="text"
+                    placeholder="Email"
+                    class="form-control"
+                    >
+                    <input
+                    required
+                    name="password"
+                    id="password"
+                    type="password"
+                    placeholder="Password"
+                    class="form-control"
+                    >
+                    <div>
+                        <button class="login-button" id="submit">Login</button>
+                    </div>
+                </form>
+            </div>
+        
+        <div id="sign-up">
+            <a href="{{url_for('registration')}}">Dont have an account? Sign up!</a>
+
+        </div>
+    </div>
+
+    </div>
+</body>
+</html>
