+++ templates/profile.html
@@ -1,0 +1,51 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link href="{{ url_for('static', filename='styles.css') }}" rel="stylesheet">
+    <title>Document</title>
+</head>
+<body>
+    <div class="container">
+        <header class="header">
+            <h1>Paws & Portions – Precision feeding with love</h1>
+        </header>
+
+        <div class="navbar">
+            <div class="nav-links">
+                <a href="{{url_for('main')}}">Home</a>
+                <a href="{{url_for('about')}}">About Us</a>
+                <a href="{{url_for('schedule')}}">Schedule</a>
+            </div>
+
+            <div class="user-profile">
+                {% if username %}
+                    <div class="profile-info">
+                        {% if user[2] %}
+                            <img src="{{ url_for('static', filename='uploads/' + user[2]) }}" alt="Profile Picture" class="profile-img">
+                        {% endif %}
+                        <a href="{{url_for('profile')}}" class="username">{{ username }}</a>
+                        <a href="{{url_for('logout')}}" class="logout">Logout</a>
+                    </div>
+                {% else %}
+                    <a href="{{url_for('login')}}" class="login-link">Login</a>
+                {% endif %}
+            </div>
+        </div>
+
+        <div class="user-info-container">
+            <div class="user-info">
+                <img src="{{ url_for('static', filename='uploads/' + user[2]) }}" alt="Profile Picture" class="profile-page-img">
+                {% if user[1] %}
+                    <p class="bio">Bio: {{ user[1] }}</p>
+                {% endif %}
+                {% if user[3] %}
+                    <p>Joined on: {{ user[3] }}</p>
+                    <a href="{{ url_for('edit_profile') }}" class="edit-profile-link">Edit Profile</a>
+                {% endif %}
+            </div>
+        </div>
+    </div>
+</body>
+</html>
